pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "syracuse"
version = "0.1.0"
description = "Syracuse/Collatz sequence exploration tools."
requires-python = ">=3.11"
dependencies = ["gudhi", "matplotlib", "numpy", "pillow", "scipy"]

[project.optional-dependencies]
dev = ["pytest"]

[project.scripts]
syracuse-generate = "syracuse.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
